Finished main page template. Created Event entity, repo, svc and controller. Created POST request for new event. Fixed field types and default values in entities

+++ backend/src/main/java/com.betanet/repository/UserRepository.java
@@ -8,5 +8,5 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Repository
 @Transactional(readOnly = true)
-public interface UserRepository extends CrudRepository<User, Long>, QuerydslPredicateExecutor<User> {
+public interface UserRepository extends CrudRepository<User, Integer>, QuerydslPredicateExecutor<User> {
 }
